Add helper functions in extensions
